🚚 네이밍을 명확하게 수정한다
