+++ src/features/record/index.ts
@@ -1,4 +1,4 @@
 export * from './models/recording-store';
 
 export * from './ui/clear-events-button';
-export * from './ui/toggle-record-button';
+export * from './ui/toggle-recording-button';
